move weather model files into own folder
